+++ src/script/bindings/mod.rs
@@ -3,6 +3,7 @@
 pub mod dns;
 pub mod helpers;
 pub mod http;
+pub mod ntp;
 pub mod ping;
 pub mod tcp;
 pub mod tls;