+++ src/script/engine.rs
@@ -41,6 +41,7 @@ pub fn build_engine(defaults: &ScriptDefaults) -> rhai::Engine {
     super::bindings::helpers::register(&mut engine);
     super::bindings::dns::register(&mut engine);
     super::bindings::http::register(&mut engine, defaults.clone());
+    super::bindings::ntp::register(&mut engine, defaults.clone());
     super::bindings::ping::register(&mut engine, defaults.clone());
     super::bindings::tcp::register(&mut engine, defaults.clone());
     super::bindings::tls::register(&mut engine);